v5-core/alert-ingestion-service/requirements.txt
fastapi
uvicorn
pydantic
kafka-python
